renaming `quiz-helpers` to `helpers`
