+++ resources/assets/components/Quiz/helpers.js
